pyproject.toml
[project]
name = "oledwall"
version = "0.1.0"
description = "OLED-friendly fuzzy circle wallpaper generator with CLI and slideshow review"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "OledWallpaperGenerator contributors"}
]
keywords = ["wallpaper", "oled", "generator", "cli", "art", "blur", "gradient"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "pyside6>=6.8.0",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
oledwall = "oledwall.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oledwall"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/presets",
    "/README.md",
    "/DESIGN.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src-path = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "SIM", "B", "C4"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
